Move training to separate module and clean up policy generation
